Moved the creation logic from the static factory method into the factory class without losing functionality
